Move helper function `getEcdsaKeyTypes` into test.
